임시 커밋

+++ be/src/main/java/com/cryptory/be/admin/dto/coin/CoinListResponseDto.java
@@ -21,7 +21,7 @@ import lombok.extern.jackson.Jacksonized;
 @AllArgsConstructor
 @Jacksonized
 public class CoinListResponseDto {
-    private Integer cryptoId;
+    private Long cryptoId;
     private String name;
     private String symbol;
     private String logoUrl;
